Checkout page added.
letsgoooooo

+++ KinoApp/KinoAppDB/Entities/TicketEntity.cs
@@ -3,7 +3,7 @@
 public class TicketEntity
 {
     public long Id { get; set; }
-    public short Status { get; set; }           // FK to LUT if you add it
+    public int Status { get; set; }           // FK to LUT if you add it
 
     public long VorstellungId { get; set; }     // show
     // public VorstellungEntity Vorstellung { get; set; } = null!; // add later when you have it
